requirements.txt
# Core dependencies
python-dotenv>=1.2.1
requests>=2.32.5

# LangChain & LangGraph core
langchain-core>=1.0.1
langchain>=1.0.2
langgraph>=1.0.1

# LLM Providers (install as needed)
# For local development with Ollama
langchain-ollama>=1.0.0

# For OpenAI (cloud)
langchain-openai>=1.0.1

# For Google Gemini (cloud)
langchain-google-genai>=3.0.0

# For Anthropic Claude (cloud)
langchain-anthropic>=1.0.0

# Knowledge Graph
neo4j>=6.0.2
langchain-community>=0.4.1

# WebSocket support for Unity bridge
websocket-client>=1.9.0

# Development & Testing
pytest>=8.4.2
pytest-asyncio>=1.2.0
